Authenticated registry mirrors: handle prefix correctly
Pulling from a prefixed and auuthenticated registry mirror
didn't work in the past.

Let's assume the following configuration is in place:

  * registry.kube.lan is an authenticated registry using
    bearer tokens as authentication method
  * A mirror of quay.io/coreos/etcd is available at
    registry.kube.lan/quay.io/coreos/etcd
  * The docker engine has the following configuration:
    ```json
      "registries": [
         {
           "Prefix" : "quay.io",
           "Mirrors": [
              {
                "URL": "https://registry.kube.lan/quay.io"
              }
           ]
         }
      ]
    ```

With this setup in place doing:

```
docker pull quay.io/coreos/etcd:v3.1
```

Should result in:

  * Have the docker engine authenticate against `registry.kube.lan`
  * Have the authenticated docker engine pull
    `registry.kube.lan/quay.io/coreos/etcd:v3.1` without the user
    being aware of it.

The docker engine checks if a registry is authenticated by doing a
`GET` request against the `/v2/` endpoint. Without this patch the docker
engine made a request against `/prefix/v2/`; in this case it would issue
the `GET` request against `https://registry.kube.lan/quay.io/v2/`.
That results in a `404` error code, which is interpreted by the engine
as "the remote registry is not using authentication, plus it doesn't
support v2".

After the docker engine is aware of the registry being authentcated, it
will contact the authentication endpoint specified by the registry.
While doing that the docker engine will provide the following
information to the authentication service:

  * `scope`: this is the name of the registry. This patch ensures the right
    name of the registry is sent. In this case it would send `registry.kube.lan`
    instead of `registry.kube.lan/quay.io`.
  * `service`: the scope of the operatation. This patch ensures the `scope`
    sent to the authentication service uses the right repository. In this case
    it would send `repository:quay.io/coreos/etcd:pull` instead of
    `repository:coreos/etcd:pull`.

Sending a wrong `service` and `scope` would cause the authentication
service to issue a wrong bearer token (something that doesn't grant any
kind of privileges or grants them against resources different from the
ones available on the registry). Without this patch the registry would
issue a `404` code to the engine, causing it to move to the next mirror.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>